Raise abstract PHP parser to feature level 8.0
