+++ src/test/php/PDepend/Source/Language/PHP/Features/PHP80/MatchExpressionTest.php
@@ -54,10 +54,9 @@ use PDepend\Source\AST\ASTVariable;
 use PDepend\Source\Parser\UnexpectedTokenException;
 
 /**
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion80
+ * @covers \PDepend\Source\Language\PHP\AbstractPHPParser
  * @copyright 2008-2017 Manuel Pichler. All rights reserved.
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
  * @group unittest
  * @group php8
  */
