+++ src/test/php/PDepend/Source/Language/PHP/PHPParserVersion80Test.php
@@ -51,12 +51,11 @@ use PDepend\Source\Tokenizer\Tokenizer;
 use PDepend\Util\Cache\CacheDriver;
 
 /**
- * Test case for the {@link \PDepend\Source\Language\PHP\PHPParserVersion80} class.
+ * Test case for the {@link \PDepend\Source\Language\PHP\AbstractPHPParser} class.
  *
- * @covers \PDepend\Source\Language\PHP\PHPParserVersion80
+ * @covers \PDepend\Source\Language\PHP\AbstractPHPParser
  * @copyright 2008-2017 Manuel Pichler. All rights reserved.
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
- *
  * @group unittest
  */
 class PHPParserVersion80Test extends AbstractTestCase
@@ -181,7 +180,7 @@ class PHPParserVersion80Test extends AbstractTestCase
     protected function createPHPParser(Tokenizer $tokenizer, Builder $builder, CacheDriver $cache)
     {
         return $this->getAbstractClassMock(
-            'PDepend\\Source\\Language\\PHP\\PHPParserVersion80',
+            'PDepend\\Source\\Language\\PHP\\PHPParserVersion81',
             [$tokenizer, $builder, $cache]
         );
     }
